Mostly cleanup.
The only truly substantive change is that a WikiTreeProfile object always extends a JSONObject which contains the profile as its top-level JSON object.
In the last version, a WikiTreeProfile would sometimes extend a JSONObject who's top-level object was the profile and sometimes extend a JSONObject who's top-level object contained the profile within it (more than a tad clunky for the user of the instance if they wanted to grab something out of the JSONObject).