Implemented a WikiTreeId class to represent a WikiTree ID. Methods which used to accept a string containing a WikiTree ID generally all require a WikiTreeId instance now.
Many methods which used to return a value or a null now return an Optional containing the value or the null (should make NPE's less likely).

Still working on having better example programs and such (a fair ways to go yet).

The key wrapped methods now come in a version that takes a WikiTreeId instance and another version that takes a long Person.Id or Space.Id.